requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
av==14.4.0
azure-ai-agents==1.2.0b4
azure-ai-projects==1.1.0b4
azure-common==1.1.28
azure-core==1.35.1
azure-identity==1.25.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
black==25.1.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudevents==1.12.0
cryptography==46.0.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
flake8==7.3.0
frozenlist==1.7.0
google-crc32c==1.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.28
lazy-object-proxy==1.12.0
MarkupSafe==3.0.2
mccabe==0.7.0
more-itertools==10.8.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.3
openai==1.107.3
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
parse==1.20.2
pathable==0.4.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
prance==25.4.8.0
propcache==0.3.2
protobuf==6.32.1
pybars4==0.9.13
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.3.0
pytest==8.4.2
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scipy==1.16.2
semantic-kernel==1.37.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
uvicorn
fastapi
bs4
